# swagger.yaml
openapi: 3.0.1
info:
  title: 영화 검색 API
  description: OMDb API를 사용하여 영화를 검색하거나 세부정보를 가져오는 API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /searchMovies:
    get:
      summary: 영화 검색
      description: 제목과 페이지를 사용하여 영화 목록을 검색합니다. OMDb API에서 해당 영화를 조회합니다.
      parameters:
        - in: query
          name: title
          schema:
            type: string
          required: true
          description: 검색할 영화의 제목 (영문으로만 검색 가능)
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          required: false
          description: 검색 결과의 페이지 번호
      responses:
        '200':
          description: 영화 검색 결과를 성공적으로 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieSearchResponse'
              examples:
                SearchExample:
                  summary: 영화 검색 결과 예시
                  value:
                    Response: 'True'
                    Search:
                      - Title: 'Inception'
                        Year: '2010'
                        imdbID: 'tt1375666'
                        Type: 'movie'
                        Poster: 'https://example.com/poster.jpg'
                    totalResults: '100'
        '404':
          description: 영화가 검색되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: '검색 된 영화가 없습니다!'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: '데이터를 처리하는 동안 오류가 발생했습니다.'

  /getMovieDetails:
    get:
      summary: 영화 세부정보 조회
      description: 영화 ID를 사용하여 특정 영화의 세부 정보를 가져옵니다.
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
          description: 세부 정보를 조회할 영화의 고유 IMDb ID
      responses:
        '200':
          description: 영화 세부 정보를 성공적으로 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetailsResponse'
              examples:
                DetailExample:
                  summary: 영화 세부 정보 조회 결과 예시
                  value:
                    Response: 'True'
                    Title: 'Inception'
                    Year: '2010'
                    imdbID: 'tt1375666'
                    Type: 'movie'
                    Poster: 'https://example.com/poster.jpg'
                    Released: '2010-07-16'
                    Runtime: '148 min'
                    Country: 'USA'
                    Plot: 'A thief who steals corporate secrets through the use of dream-sharing technology is given a final chance at redemption...'
                    Ratings:
                      - Source: 'Internet Movie Database'
                        Value: '8.8/10'
                      - Source: 'Rotten Tomatoes'
                        Value: '87%'
                    Actors: 'Leonardo DiCaprio, Joseph Gordon-Levitt, Ellen Page'
                    Director: 'Christopher Nolan'
                    Production: 'Warner Bros.'
                    Genre: 'Action, Adventure, Sci-Fi'
        '404':
          description: 영화가 검색되지 않음
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: '검색 된 영화가 없습니다!'
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: '데이터를 처리하는 동안 오류가 발생했습니다.'

components:
  schemas:
    MovieSearchResponse:
      type: object
      properties:
        Response:
          type: string
          example: 'True'
        Search:
          type: array
          items:
            type: object
            properties:
              Title:
                type: string
                example: 'Inception'
              Year:
                type: string
                example: '2010'
              imdbID:
                type: string
                example: 'tt1375666'
              Type:
                type: string
                example: 'movie'
              Poster:
                type: string
                example: 'https://example.com/poster.jpg'
        totalResults:
          type: string
          example: '100'
        Error:
          type: string
          example: ''

    MovieDetailsResponse:
      type: object
      properties:
        Response:
          type: string
          example: 'True'
        Title:
          type: string
          example: 'Inception'
        Year:
          type: string
          example: '2010'
        imdbID:
          type: string
          example: 'tt1375666'
        Type:
          type: string
          example: 'movie'
        Poster:
          type: string
          example: 'https://example.com/poster.jpg'
        Released:
          type: string
          example: '2010-07-16'
        Runtime:
          type: string
          example: '148 min'
        Country:
          type: string
          example: 'USA'
        Plot:
          type: string
          example: 'A thief who steals corporate secrets through the use of dream-sharing technology is given a final chance at redemption...'
        Ratings:
          type: array
          items:
            type: object
            properties:
              Source:
                type: string
                example: 'Internet Movie Database'
              Value:
                type: string
                example: '8.8/10'
        Actors:
          type: string
          example: 'Leonardo DiCaprio, Joseph Gordon-Levitt, Ellen Page'
        Director:
          type: string
          example: 'Christopher Nolan'
        Production:
          type: string
          example: 'Warner Bros.'
        Genre:
          type: string
          example: 'Action, Adventure, Sci-Fi'